requirements.txt
requests
python-dotenv
pandas
numpy
kafka-python
pyspark==3.5.1
matplotlib
plotly
notebook
ipykernel
retrying